requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
flask-cors>=3.0.0
scapy>=2.4.0
watchdog>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
python-whois>=0.7.0
